#28 [MOD] : calculateTime 통합
